refactor: 관심사에 따라, validation, type, constants로 분리
