+++ src/AddCard/components/AddCardPreview/constants/index.ts
@@ -1,7 +1,7 @@
-import type { CardBrandImages } from "../types";
 import VisaCardImage from "@assets/icons/visa-card.svg";
 import MasterCardImage from "@assets/icons/master-card.svg";
-export const CARD_BRAND_IMAGES: CardBrandImages = {
+import { OverseaCardBrand } from "../types";
+export const CARD_BRAND_IMAGES: Record<OverseaCardBrand, string> = {
   VISA: VisaCardImage,
   MASTERCARD: MasterCardImage,
   DEFAULT: "",
