+++ src/AddCard/constants/index.ts
@@ -1,0 +1,9 @@
+import type { FlowStep } from "../types/hook";
+export const STEP_ORDER: FlowStep[] = [
+  "CARD_NUMBER",
+  "CARD_BRAND",
+  "EXPIRE_DATE",
+  "CVC",
+  "PASSWORD",
+  "COMPLETE",
+];
